+++ SpinIDE/OpenSpin/preprocess.cpp
@@ -585,7 +585,7 @@ expand_macros(struct preprocess *pp, struct flexbuf *dst, char *src)
         }
         flexbuf_addstr(dst, def);
     }
-    len = flexbuf_curlen(dst);
+    len = (int) flexbuf_curlen(dst);
     flexbuf_addchar(dst, 0);
     return len;
 }
